Jobs for pets

+++ src/main/kotlin/Effects.kt
@@ -4,7 +4,7 @@ import org.jetbrains.exposed.sql.Table
 object Effects : Table("effects") {
     val id: Column<Int> = integer("id")
     val action: Column<Int> = integer("action")
-    val params: Column<List<Int>> = array<Int>("params")
+    val params: Column<List<Float>> = array<Float>("params")
 
     override val primaryKey = PrimaryKey(id)
 }
